Functional Backend Update:
Files are all working with Retrofit Client Instance now. We just need to establish a base url JSON file and we should be able to setup data for our app very soon! Once a base url is in place, we can begin to instantiate our JSON in the front end.